pyproject.toml
[project]
name = "trr-scripts"
version = "0.1.0"
description = "Smart contract analysis scripts for Ethereum/Arbitrum"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "slither-analyzer>=0.10.0",
    "networkx>=3.0",
    "numpy>=1.24.0",
    "web3>=6.0.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
trr = "trr:main"


[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["trr.py"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"test_*.py" = ["E501"]    # Allow long lines in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

